Update to latest.

+++ scriptmodules/emulators/zesarux.sh
@@ -12,14 +12,14 @@
 rp_module_id="zesarux"
 rp_module_desc="ZX Spectrum emulator ZEsarUX"
 rp_module_menus="4+"
-rp_module_flags="dispmanx"
+rp_module_flags="dispmanx !x86"
 
 function depends_zesarux() {
     getDepends libssl-dev libpthread-stubs0-dev libsdl1.2-dev libasound2-dev
 }
 
 function sources_zesarux() {
-    wget -O- -q "http://downloads.petrockblock.com/retropiearchives/ZEsarUX_src-3.0.tar.gz" | tar -xvz --strip-components=1
+    wget -O- -q "$__archive_url/ZEsarUX_src-3.0.tar.gz" | tar -xvz --strip-components=1
 }
 
 function build_zesarux() {
@@ -53,9 +53,9 @@ _EOF_
     chmod +x "$romdir/zxspectrum/+Start ZEsarUX.sh"
     chown $user:$user "$romdir/zxspectrum/+Start ZEsarUX.sh"
 
-    ln -sf "$configdir/zxspectrum/.zesaruxrc" "$home/.zesaruxrc"
-
-    cat > "$configdir/zxspectrum/.zesaruxrc" << _EOF_
+    moveConfigFile "$home/.zesaruxrc" "$configdir/zxspectrum/.zesaruxrc"
+    if [[ ! -f "$configdir/zxspectrum/.zesaruxrc" ]]; then
+        cat > "$configdir/zxspectrum/.zesaruxrc" << _EOF_
 ;ZEsarUX sample configuration file
 ;
 ;Lines beginning with ; or # are ignored
@@ -74,8 +74,8 @@ _EOF_
 ;Remap Fire Event. Uncomment and amend if you wish to change the default button 3.
 ;--joystickevent 3 Fire
 _EOF_
-
-    chown $user:$user "$configdir/zxspectrum/.zesaruxrc"
+        chown $user:$user "$configdir/zxspectrum/.zesaruxrc"
+    fi
 
     setDispmanx "$md_id" 1
